requirements.txt
python-telegram-bot
edge-tts
SpeechRecognition
pydub
ffmpeg-python
fastapi
uvicorn
